requirements.txt
Flask
PyMySQL
python-dotenv
